animeshki-backend/src/display-service/requirements.txt
uvicorn
fastapi
jinja2
python-dotenv
pyjwt
bcrypt